extract validation and exec logic from main
