+++ app/backend/input_validation.py
@@ -1,0 +1,66 @@
+from uuid import UUID
+from pathlib import Path
+"""
+Summary:
+    Validates user's input path. Input path must be a valid filepath. Cannot contain files more then 4GB and must not be a rar file (zip file allowed).
+Params: 
+    - User inputted filepath as string
+Return:
+"""
+def validate_path(filepath: str) -> Path:
+    max_size_bytes: int = 4 * 1024 * 1024 * 1024  # 4gb limit
+
+    def _is_rar_file(path: Path) -> bool:
+        return path.suffix.lower() in ['.rar', '.r00', '.r01']
+    
+    def _get_directory_size(path: Path) -> int:
+        total: int = 0
+        try:
+            for file_path in path.rglob('*'):
+                if file_path.is_file():
+                    try:
+                        total += file_path.stat().st_size
+                    except (OSError, PermissionError) as e:
+                        print(f"Cannot access file {file_path}: {e}")
+                        continue
+        except (OSError, PermissionError) as e:
+            raise ValueError(f"Cannot access directory: {e}")
+        return total
+
+    filepath = filepath.strip().strip('"').strip("'")
+    if not filepath:
+        raise ValueError("Filepath cannot be empty")
+    
+    try:
+        path: Path = Path(filepath).expanduser().resolve()
+    except (OSError, RuntimeError) as e:
+        raise ValueError(f"Invalid file path: {e}")
+
+    if not path.exists():
+        raise FileNotFoundError(f"Path not found: {filepath}")
+    
+    if path.is_file() and _is_rar_file(path):
+        raise ValueError(f"RAR files are not supported: {filepath}")
+    
+    if path.is_file():
+        try:
+            size: int = path.stat().st_size
+            if size > max_size_bytes:
+                size_gb: float = size/(1024 ** 3)
+                raise ValueError(f"File too large: {size_gb:.2f}GB (max 4GB)")
+        except (OSError, PermissionError) as e:
+            raise ValueError(f"Cannot access file: {e}")
+
+    elif path.is_dir():
+        total_size: int = _get_directory_size(path)
+        if total_size > max_size_bytes:
+            size_gb: float = total_size / (1024 ** 3)
+            raise ValueError(f"Folder too large: {size_gb:.2f}GB (max 4GB)")   
+    return path
+
+def validate_uuid(uuid:str)-> str:
+    result = UUID(uuid) #Try creating an UUID obj, if it fails we know its either wrong version or invalid
+    if result:
+        return uuid
+    else:
+        raise ValueError('Invalid UUID')
